pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "bess"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyomo",
    "highspy",
    "pandas",
    "requests",
    "python-dotenv",
    "pyarrow",
    "lightgbm",
    "scikit-learn",
    "streamlit==1.43.0",
    "altair==4.2.2",
    "plotly",
    "pytest",
    "matplotlib",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]